Fix EBS attachment check always evaluating true.

+++ elastic-beanstalk/.platform/hooks/predeploy/01-mount-ebs.sh
@@ -9,23 +9,32 @@ INSTANCE_ID=$(curl -H "X-aws-ec2-metadata-token: $IMDS_TOKEN" -s  http://169.254
 DATA_VOLUME_ID=$(/opt/elasticbeanstalk/bin/get-config environment -k DATA_VOLUME_ID)
 
 # Check if volume is already attached.
-if DATA_VOLUME_ATTACHED=$(aws ec2 describe-volumes --filters Name=attachment.instance-id,Values=$INSTANCE_ID | jq --arg DATA_VOLUME_ID "$DATA_VOLUME_ID" 'contains({Volumes: [{VolumeId: $DATA_VOLUME_ID}]})')
+VOLUME_ATTACHED=$(aws ec2 describe-volumes --filters Name=attachment.instance-id,Values=$INSTANCE_ID | jq --arg DATA_VOLUME_ID $DATA_VOLUME_ID 'contains({"Volumes": [{"VolumeId": $DATA_VOLUME_ID}]})')
+echo "Volume attached evaluated to $VOLUME_ATTACHED"
+if [ VOLUME_ATTACHED = "true" ]
 then
-    echo "Volume '${DATA_VOLUME_ID}' already attached to Instance '${INSTANCE_ID}'"
-    exit
+    echo "Volume '$DATA_VOLUME_ID' already attached to Instance '$INSTANCE_ID'"
 else
-echo "Volume '${DATA_VOLUME_ID}' not attached to current instance."
-echo "Attempt to attach to self."
+    echo "Volume '$DATA_VOLUME_ID' not attached to current instance."
+    echo "Attempt to attach to self."
 
-# Wait for the volume to be available
-#   Currently can't handle if the volume is attached to an instance
-#   other than the current application.
-aws ec2 wait volume-in-use --region us-east-2 --volume-ids $DATA_VOLUME_ID
+    # Wait for the volume to be available
+    #   Currently can't handle if the volume is attached to an instance
+    #   other than the current application.
+    aws ec2 wait volume-in-use --region us-east-2 --volume-ids $DATA_VOLUME_ID
 
-# Attach volume to this application instance on /dev/sdh
-aws ec2 attach-volume --volume-id $DATA_VOLUME_ID --instance-id $INSTANCE_ID --device /dev/sdh
-
-mkfs -t ext4 /dev/sdh
-mkdir /data/mysql
-mount /dev/sdh/mysql /data/mysql
+    # Attach volume to this application instance on /dev/sdh
+    aws ec2 attach-volume --volume-id $DATA_VOLUME_ID --instance-id $INSTANCE_ID --device /dev/sdh
 fi
+
+# Create filesystem on volume if it's empty
+if [ $(blkid | grep -c "/dev/sdh") -ne 0 ]
+then
+    echo "Volume is empty, creating filesystem..."
+    mkfs -t ext4 /dev/sdh
+fi
+
+# Mount EBS to folder for MySQL docker image
+echo "Mounting EBS device at /dev/sdh to /data/mysql if it isn't already..."
+mkdir -p /data/mysql
+mount /dev/sdh/mysql /data/mysql
